% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{clear}
\alias{clear}
\title{Clear current query}
\usage{
clear()
}
\value{
TRUE (invisible)
}
\description{
Clear current query
}
\seealso{
\code{\link[=query]{query()}}
for opening a query

\code{\link[=submit]{submit()}}
for submitting a query
}
