% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Create a query}
\usage{
query(
  q = call("member", expression(X), list(quote(a), "b", 3L, 4, TRUE, expression(Y)))
)
}
\arguments{
\item{q}{an R call. The R call consists of symbols, integers and real numbers,
character strings, boolean values, expressions, lists, and other calls.
Vectors of booleans, integers, floating point numbers, and strings with
length \emph{N} > 1 are translated to prolog compounds !/N, \%/N, #/N and $$/N,
respectively.}
}
\value{
If the creation of the query succeeds, \code{TRUE}
}
\description{
Create a query
}
\details{
SWI-Prolog does not allow multiple open queries. If another query is open, it
it is closed and a warning is shown.
}
\examples{
q <- query(call("member", 1, expression(X)))
submit()
submit()
clear()

}
\seealso{
\code{\link[=submit]{submit()}}
for submitting a query

\code{\link[=clear]{clear()}}
to close the currently open query
}
