% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswipl.R
\name{rswipl_init}
\alias{rswipl_init}
\title{Start prolog}
\usage{
rswipl_init(argv1 = commandArgs()[1])
}
\arguments{
\item{argv1}{file name of the R executable}
}
\value{
`TRUE` on success
}
\description{
Start prolog
}
\details{
SWI-Prolog is automatically initialized when the rswipl library is loaded, so
this function is normally not directly invoked.
}
