% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{submit}
\alias{submit}
\title{Submit a query that has been opened with \code{\link[=query]{query()}} before.}
\usage{
submit()
}
\value{
If the query fails, \code{FALSE} is returned. If the query succeeds, a
(possibly empty) list is returned that includes the bindings required to
satisfy the query.
}
\description{
Submit a query that has been opened with \code{\link[=query]{query()}} before.
}
\seealso{
\code{\link[=query]{query()}}
for a opening a query.

\code{\link[=clear]{clear()}}
for a clearing a query.
}
