% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswipl.R
\name{swipl}
\alias{swipl}
\title{Invoke SWI-Prolog}
\usage{
swipl(sigalert = NA)
}
\arguments{
\item{sigalert}{Use a different alert signal than SIGUSR2 (ignored on Windows)}
}
\value{
nothing useful
}
\description{
This function is internally used to emulate swipl -g goal using
the R program: R -e "rswipl::swipl()" -q --no-echo --args -g goal
}
